Create LICENSE.md
In this example, we are generating a random phrase using a set of words from the nouns, verbs, and adjectives arrays. We select a random word from each array using the Random random number generator. Then we combine the selected words together and display the resulting phrase on the screen. Each time you run the program, a new unique phrase will be generated.